+++ packages/sdks/web-component/test/descope-wc.test-harness.ts
@@ -1,7 +1,11 @@
 /* eslint-disable max-classes-per-file */
 // @ts-nocheck
 
-import { createSdk, ensureFingerprintIds } from '@descope/web-js-sdk';
+import {
+  createSdk,
+  ensureFingerprintIds,
+  getSessionToken,
+} from '@descope/web-js-sdk';
 import '@testing-library/jest-dom';
 import { getABTestingKey } from '../src/lib/helpers/abTestingKey';
 import { resetCustomStorage } from '../src/lib/helpers/storage';
@@ -14,9 +18,12 @@ jest.mock('@descope/web-js-sdk', () => ({
   createSdk: jest.fn(),
   clearFingerprintData: jest.fn(),
   ensureFingerprintIds: jest.fn(),
+  getSessionToken: jest.fn(),
 }));
 
 export { createSdk, ensureFingerprintIds };
+
+export const getSessionTokenMock = getSessionToken as jest.Mock;
 
 export const WAIT_TIMEOUT = 25000;