+++ packages/sdks/web-js-sdk/src/index.ts
@@ -40,7 +40,11 @@ export { getSessionToken } from './enhancers/withPersistTokens/helpers';
 
 export { hasOidcParamsInUrl } from './sdk/oidc/helpers';
 
-export type { JWTResponse } from '@descope/core-js-sdk';
+export type {
+  JWTResponse,
+  FlowStartOptions,
+  FlowNextOptions,
+} from '@descope/core-js-sdk';
 export type { OneTapConfig } from './sdk/fedcm';
 export type { CookieConfig } from './enhancers/withPersistTokens/types';
 export type { FlowNonceOptions } from './enhancers/withFlowNonce/types';