+++ packages/widgets/outbound-applications-widget/CHANGELOG.md
@@ -2,6 +2,31 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.6.2](https://github.com/descope/descope-js/compare/outbound-applications-widget-0.6.1...outbound-applications-widget-0.6.2) (2026-09-08)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.52.1`
+* `web-component` updated to version `4.3.2`
+## [0.6.1](https://github.com/descope/descope-js/compare/outbound-applications-widget-0.6.0...outbound-applications-widget-0.6.1) (2026-09-07)
+
+### Dependency Updates
+
+* `sdk-mixins` updated to version `0.26.1`
+* `sdk-component-drivers` updated to version `0.16.0`
+* `web-component` updated to version `4.3.1`
+## [0.6.0](https://github.com/descope/descope-js/compare/outbound-applications-widget-0.5.14...outbound-applications-widget-0.6.0) (2026-08-24)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.52.0`
+* `sdk-mixins` updated to version `0.26.0`
+* `web-component` updated to version `4.3.0`
+
+### Features
+
+* **widgets:** componentsConditionsMixin to apply server condition verdicts RELEASE ([#1463](https://github.com/descope/descope-js/issues/1463)) ([68634d1](https://github.com/descope/descope-js/commit/68634d17097e9687c96d088ab1ad7cecdd889cbb)), closes [descope/etc#17289](https://github.com/descope/etc/issues/17289)
+
 ## [0.5.14](https://github.com/descope/descope-js/compare/outbound-applications-widget-0.5.13...outbound-applications-widget-0.5.14) (2026-08-17)
 
 ### Dependency Updates